% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{run_apti_app}
\alias{run_apti_app}
\title{Run APTI 'Shiny' App}
\usage{
run_apti_app()
}
\value{
Launches a Shiny application.
}
\description{
Launches a 'Shiny' app for interactive APTI calculation and visualization.
}
\examples{
if (interactive()) {
  run_apti_app()
}
}
