% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data for APTI calculation}
\format{
A data frame with 15 rows and 4 variables:
\describe{
  \item{A}{Ascorbic acid content (mg/g)}
  \item{TC}{Total chlorophyll content (mg/g)}
  \item{P}{Leaf extract pH}
  \item{R}{Relative water content (\%)}
}
}
\source{
Manually entered for demonstration
}
\usage{
sample_data
}
\description{
A dataset containing biochemical parameters used to compute APTI.
}
\keyword{datasets}
